break project into modules
