+++ beam-sdks-java-io-solace/src/main/java/com/solace/apache/beam/StringMessageMapper.java
@@ -1,0 +1,14 @@
+package com.solace.apache.beam;
+
+import com.solacesystems.jcsmp.BytesXMLMessage;
+
+import java.nio.charset.StandardCharsets;
+
+public class StringMessageMapper implements SolaceIO.MessageMapper<String> {
+  private static final long serialVersionUID = 42L;
+
+  @Override
+  public String mapMessage(BytesXMLMessage message) throws Exception {
+    return new String(message.getBytes(), StandardCharsets.UTF_8);
+  }
+}
