+++ pom.xml
@@ -20,9 +20,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.example</groupId>
-  <artifactId>beam-solaceio</artifactId>
-  <version>0.1</version>
+  <groupId>com.solace.apache.beam</groupId>
+  <artifactId>solace-apache-beam-parent</artifactId>
+  <version>1.0.0-SNAPSHOT</version>
 
   <packaging>jar</packaging>
 
@@ -61,6 +61,11 @@
       </snapshots>
     </repository>
   </repositories>
+
+  <modules>
+    <module>beam-sdks-java-io-solace</module>
+    <module>solace-apache-beam-samples</module>
+  </modules>
 
   <build>
     <plugins>
@@ -426,7 +431,7 @@
       <artifactId>hamcrest-core</artifactId>
       <version>${hamcrest.version}</version>
     </dependency>
-    
+
     <dependency>
       <groupId>org.hamcrest</groupId>
       <artifactId>hamcrest-library</artifactId>
@@ -446,7 +451,7 @@
       <version>${beam.version}</version>
       <scope>test</scope>
     </dependency>
-    
+
     <dependency>
       <groupId>org.mockito</groupId>
       <artifactId>mockito-core</artifactId>
