+++ solace-apache-beam-samples/src/main/resources/log4j.properties
