+++ solace-apache-beam-samples/src/main/resources/simplelogger.properties
